requirements.txt
matplotlib==3.4.2
pyyaml==5.4.1
opencv-python==4.5.1.48
easydict==1.9
scikit-image==0.17.2
tensorboardx==2.1
tensorboard==1.15.0
lpips==0.1.3
fsspec==2022.7.1

tqdm==4.60.0
transformers==4.24.0
torchmetrics==0.7.3

einops==0.3.0
omegaconf==2.1.1
open_clip_torch==2.0.2
webdataset==0.2.5
huggingface-hub==0.11.1
gradio==3.17.1
